requirements.txt
amqp==5.1.1
billiard==4.1.0
black==23.9.1
blinker==1.6.2
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
six==1.16.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
